Inline logs bundle handling and split start/stop bundle files
- Inline BundleLogsOverride directly into logs.go since it's a custom command
- Split start_stop_bundle.go into separate start_bundle.go and stop_bundle.go
- Split corresponding test files for better organization
- Update acceptance test output for "Deploying project..." message
