+++ acceptance/apps/deploy/bundle-no-args/output.txt
@@ -1,6 +1,6 @@
 
 >>> [CLI] apps deploy --skip-validation
-Deploying bundle...
+Deploying project...
 Uploading bundle files to /Workspace/Users/[USERNAME]/.bundle/test-bundle/default/files...
 Deploying resources...
 Updating deployment state...
