+++ cmd/apps/stop_bundle_test.go
@@ -8,22 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestBundleStartOverrideWithWrapper(t *testing.T) {
-	mockWrapper := func(cmd *cobra.Command, appName string, err error) error {
-		return err
-	}
-
-	overrideFunc := BundleStartOverrideWithWrapper(mockWrapper)
-	assert.NotNil(t, overrideFunc)
-
-	cmd := &cobra.Command{}
-	startReq := &apps.StartAppRequest{}
-
-	overrideFunc(cmd, startReq)
-
-	assert.Equal(t, "start [NAME]", cmd.Use)
-}
-
 func TestBundleStopOverrideWithWrapper(t *testing.T) {
 	mockWrapper := func(cmd *cobra.Command, appName string, err error) error {
 		return err
@@ -38,22 +22,6 @@ func TestBundleStopOverrideWithWrapper(t *testing.T) {
 	overrideFunc(cmd, stopReq)
 
 	assert.Equal(t, "stop [NAME]", cmd.Use)
-}
-
-func TestBundleStartOverrideHelpText(t *testing.T) {
-	mockWrapper := func(cmd *cobra.Command, appName string, err error) error {
-		return err
-	}
-
-	cmd := &cobra.Command{}
-	startReq := &apps.StartAppRequest{}
-
-	overrideFunc := BundleStartOverrideWithWrapper(mockWrapper)
-	overrideFunc(cmd, startReq)
-
-	assert.NotEmpty(t, cmd.Long)
-	assert.Contains(t, cmd.Long, "Start an app")
-	assert.Contains(t, cmd.Long, "project directory")
 }
 
 func TestBundleStopOverrideHelpText(t *testing.T) {
